([bs]<=[*]      interlude: :✓+[#libricate] (.. 2625))
Synopsis: to libricate this badboy we pulled out
   the old search & replace, dusted it off and
   fixed a buglette in it near finding custom
   functions. along they way we jumped some
   next commit stuff to here for easier
   integration

[br] • in top file we re-arrange things to the
       new standard-ish order of sections:
       singleton methods, support modules,
       then the `Lib_`, then constants

     • in model, it now peeks for an "action.rb"
       as well as an "actions/", to see if the
       node defines actions (without needing to
       load the file) (the only reason we do this
       is because the autoloader loads whole
       directory listings at once instead of
       reading for individual files, otherwise
       we wouldn't incur the cost). although we
       ended up not needing this yet because we
       broke out actons into their own files
       after all, we are keeping this around,
       anticpating wanting this in the future when
       we have better inferred actions.

     • the original inspiration for this interlude,
       we modernized the interface for git config
       persist (use paths not pathnames; we hard-
       code logic for writing first to tempfiles
       for now (that is, it is an option; but the
       path to the tempfile is deterministic and
       effectively not [#hl-100] thread-safe,
       which again is a reminder that this whole
       house is made of cards and needs to be
       blown away once we used it for a rocketship))

     • sneak preview of the newer
       entity edit session API, to be explained
       further in the next commit

(the only remaining to libricate: [f2],[cb],[ts])

(2 in 12.5)

+++ lib/skylab/brazen/test/cli/actions/test-support.rb
@@ -96,7 +96,9 @@ module Skylab::Brazen::TestSupport::CLI::Actions
     def blow_away path
       td_path = tmpdir.to_path
       if td_path == path[ 0, td_path.length ] && path.include?( '/brAzen/' )
-        debug_IO.puts "rm -rf #{ path }"
+        if do_debug
+          debug_IO.puts "rm -rf #{ path }"
+        end
         file_utils.remove_entry_secure path  # SCARY
       else
         self._SANITY
